name changes and other requested changes
