+++ src/components/ApiDelegationOverviewPage/ApiDelegationAccordion/DeletableListItem/DeletableListItem.module.css
@@ -1,20 +1,20 @@
-.deletable-list-item__deleteSection {
+.deleteSection {
     display: flex;
     align-items: center;
 }
 
-.deletable-list-item__listItem {
+.listItem {
     display: flex;
     margin: 0.3rem;
 }
 
-.deletable-list-item__itemText {
+.itemText {
     flex: 1 1 auto;
     display: flex;
     align-items: center;
 }
 
-.deletable-list-item__itemText--soft-delete {
+.itemText__softDelete {
     text-decoration: line-through;
     opacity: 70%;
 }
